Upgrade the coffee app and move coroutines to the released, non-experiemntal versions.  Regenerate the deps.
