+++ examples/dagger/src/coffee/CoffeeApp.kt
@@ -16,7 +16,7 @@
 package coffee
 
 import dagger.Component
-import kotlinx.coroutines.experimental.runBlocking
+import kotlinx.coroutines.runBlocking
 import javax.inject.Singleton
 import tea.TeaPot
 
@@ -30,7 +30,7 @@ class CoffeeApp {
     companion object {
         @JvmStatic
         fun main(args: Array<String>) {
-            if(TeaPot.isEmpty()) {
+            if (TeaPot.isEmpty()) {
                 val coffeeShop = DaggerCoffeeApp_CoffeeShop.builder().build()
                 runBlocking {
                     coffeeShop.maker().brew()
